refactor: API 폴더 내부 구조 정리 및 미적용 파일에 일관된 구조 적용 (#80)
