--- src/api/clients/publicApi.ts
@@ -1,10 +1,0 @@
-import axios from 'axios';
-
-const publicApi = axios.create({
-  baseURL: import.meta.env.VITE_API_BASE_URL,
-  headers: {
-    'Content-Type': 'application/json',
-  },
-});
-
-export default publicApi;
